Abstract block level logic in data server to block data server handler.
